feature: axios setup + register user api hit
