+++ frontend/app/layout.tsx
@@ -1,5 +1,7 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
+
+import ReactQueryClientProvider from '@/providers/ReactQueryClientProvider';
 import './globals.css';
 
 const geistSans = Geist({
@@ -24,11 +26,13 @@ export default function RootLayout({
 }>) {
   return (
     <html lang='en'>
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {children}
-      </body>
+      <ReactQueryClientProvider>
+        <body
+          className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+        >
+          {children}
+        </body>
+      </ReactQueryClientProvider>
     </html>
   );
 }
