+++ frontend/lib/api-routes/auth.routes.ts
@@ -1,0 +1,10 @@
+import { UserProps } from '@/components/RegisterForm';
+import $axios from '../axios/axios.instance';
+
+export const registerUser = async (values: UserProps) => {
+  return await $axios.post('/auth/register', values);
+};
+
+// export const loginUser = async (values: any) => {
+//   return await $axios.post('/auth/login', values);
+// };
