+++ frontend/lib/axios/axios.instance.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+const $axios = axios.create({
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
+  timeout: 10000,
+});
+
+export default $axios;
