+++ frontend/package.json
@@ -14,6 +14,8 @@
     "@mui/icons-material": "^6.3.0",
     "@mui/material": "^6.3.0",
     "@mui/x-date-pickers": "^7.23.3",
+    "@tanstack/react-query": "^5.62.11",
+    "axios": "^1.7.9",
     "dayjs": "^1.11.13",
     "formik": "^2.4.6",
     "next": "15.1.2",
