+++ frontend/providers/ReactQueryClientProvider.tsx
@@ -1,0 +1,16 @@
+'use client';
+import React from 'react';
+
+import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
+
+const queryClient = new QueryClient();
+
+const ReactQueryClientProvider = ({
+  children,
+}: Readonly<{ children: React.ReactNode }>) => {
+  return (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+};
+
+export default ReactQueryClientProvider;
